+++ collective/jsonify/wrapper.py
@@ -1,5 +1,6 @@
 from Acquisition import aq_base
 from DateTime import DateTime
+from Products.CMFCore.utils import getToolByName
 import datetime
 import os
 
@@ -778,3 +779,31 @@ class Wrapper(dict):
             self['document_src'] = self.decode(document_src())
         else:
             self['_zopeobject_document_src'] = ''
+
+
+    def get_history(self):
+        """ Export the history - metadata
+        """
+        try:
+            repo_tool = getToolByName(self.context, "portal_repository")
+            history_metadata = repo_tool.getHistoryMetadata(self.context)
+            if not(hasattr(history_metadata,'getLength')):
+                # No history metadata
+                return
+
+            history_list = []
+            # Count backwards from most recent to least recent
+            for i in xrange(history_metadata.getLength(countPurged=False)-1, -1, -1):
+                data = history_metadata.retrieve(i, countPurged=False)
+                meta = data["metadata"]["sys_metadata"].copy()
+                version_id = history_metadata.getVersionId(i, countPurged=False)
+                try:
+                    dateaux = datetime.datetime.fromtimestamp(meta.get('timestamp',0))
+                    meta['timestamp'] = dateaux.strftime("%Y/%m/%d %H:%M:%S GMT")
+                except Exception, ex:
+                    meta['timestamp']=''
+                history_list.append(meta)
+            self['_history'] = history_list
+
+        except Exception, ex:
+            pass